Add endpoint DomainOrganic

+++ backend/app/app/models/domain_organic.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, DECIMAL
+from sqlalchemy import Column, Integer, String, DECIMAL, BigInteger
 
 from app.db.base_class import Base
 
@@ -17,5 +17,5 @@ class DomainOrganic(Base):
     url = Column(String, nullable=False)
     traffic = Column(DECIMAL, nullable=False, default=0.0)
     traffic_cost = Column(DECIMAL, nullable=False, default=0.0)
-    number_results = Column(Integer, nullable=False, default=0)
+    number_results = Column(BigInteger, nullable=False, default=0)
     trends = Column(String, index=True, nullable=False)
